chore(provisioner/terraform): extract terraform parsing logic to package tfextract
